I think i fixed server crash with custom skins.
The issue with the crash was that skins was being applied to clients that's was not in the game. Main common issue was that if minions had the skin changer this could cause the crash when minions disappear when skins applied. Small changes has been made to fix this issue.

@raphyjb